cloud_functions/requirements.txt
functions-framework==3.8.2
google-auth==2.38.0
requests==2.32.3
flask==3.1.0
